requirements/base.txt
Django==2.2.7

six==1.12.0
psycopg2-binary==2.8.2
requests==2.22.0
django-model-utils==3.2.0
django-markdown-deux==1.0.5
django-localflavor==2.2
django-extensions==2.2.5
djangorestframework==3.10.3
django-redis==4.10.0
freezegun==0.3.12
hiredis==1.0.0
redis==3.3.11
Pillow==6.2.0
icalendar==4.0.3
dealer==2.1.0
django-robots==3.1.0
django-debug-toolbar==1.11
gocardless_pro==1.9.0
newspaper3k==0.2.8

git+git://github.com/DemocracyClub/dc_base_theme.git@0.3.9
git+https://github.com/DemocracyClub/dc_signup_form.git@2.1.0
djangorestframework-jsonp==1.0.2
feedparser==5.2.1

sentry-sdk==0.13.1
sopn-publish-date==1.3.2

stripe==2.28.0
